// src/app/interfaces/gift-exchange-pairs.interface.ts
export interface GiftExchangePairs {
    pairs: string[];
    originalList: string[];
    shared: boolean;
    displayName?: string;
    id?: string;
    uid?: string;
}

export interface GiftExchangePairsJSON {
    pairs: string;
    originalList: string;
    shared: boolean;
    displayName?: string;
    id?: string;
    uid?: string;
}